extract serializers into own file and add test
